+++ lib/unicode/string/dfa/grapheme.ex
@@ -1,0 +1,4 @@
+defmodule Unicode.String.Dfa.Grapheme do
+  @moduledoc false
+  use Unicode.String.Dfa, type: "GraphemeClusterBreak"
+end