+++ lib/unicode/string/dfa/line.ex
@@ -1,0 +1,4 @@
+defmodule Unicode.String.Dfa.Line do
+  @moduledoc false
+  use Unicode.String.Dfa, type: "LineBreak", dictionary_pass: true
+end