+++ lib/unicode/string/dfa/sentence.ex
@@ -1,0 +1,4 @@
+defmodule Unicode.String.Dfa.Sentence do
+  @moduledoc false
+  use Unicode.String.Dfa, type: "SentenceBreak"
+end