+++ lib/unicode/string/dfa/word.ex
@@ -1,0 +1,4 @@
+defmodule Unicode.String.Dfa.Word do
+  @moduledoc false
+  use Unicode.String.Dfa, type: "WordBreak"
+end